+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.1.0
+
+- Added custom portal and auth cookie config options (Thanks to @Delivator)
+- Renamed config file
+
 ## 1.0.0
 
 - Initial version